add page object for additional claimants and upload
